# [wasm] optimize interop startup and linker hints