Prepare for making legacy interop optional
- move methods used only in legacy into `LegacyHostImplementation` class
- rename `GetTaskResult` to `GetTaskResultDynamic`
- make `s_taskGetResultMethodInfo` initialization lazy
- rationalize DynamicDependency 

Perf:
new FF     AppStart, Reach managed: 355.1764705882353ms
new chrome AppStart, Reach managed: 249.43478260869566ms
old FF     AppStart, Reach managed: 359.3529411764706ms
old chrome AppStart, Reach managed: 256.6ms