@pavelsavara 2022-12-08
/azp run runtime-wasm